Extract a method for locating Bazel
`bazelisk` can be thought of as two separate pieces. The first of these is a
mechanism for locating a particular `bazel` version, possibly downloading it
from the Web, or using a local install. The second piece is the additional
functionality for flag flipping and finding out where problems have occurred
within commits to the project.

This PR extracts the location of the `bazel` binary into its own method which
can be used by other tooling that also needs to locate the binary.